tools.py
import logging
from livekit.agents import function_tool , RunContext
import requests

# search engine 
from langchain_community.tools import DuckDuckGoSearchRun

import os
import smtplib
from email.mime.multipart import MIMEMultipart  
from email.mime.text import MIMEText
from typing import Optional, List , Any , Dict


from gan2.data_loader import load_all_documents

from gan2.vectorstore import FaissVectorStore

@function_tool()
async def get_weather(context: RunContext,  city: str) -> str:
    """
    Get the current weather for a given city.
    
    """
    try:
        response = requests.get(
            f"https://wttr.in/{city}?format=3")
        if response.status_code == 200:
            logging.info(f"Weather for {city}: {response.text.strip()}")
            return response.text.strip()   
        else:
            logging.error(f"Failed to get weather for {city}: {response.status_code}")
            return f"Could not retrieve weather for {city}."
    except Exception as e:
        logging.error(f"Error retrieving weather for {city}: {e}")
        return f"An error occurred while retrieving weather for {city}." 



@function_tool()
async def search_web(context: RunContext,  query: str) -> str:
    """
    Search the web using DuckDuckGo.
    """
    try:
        results = DuckDuckGoSearchRun().run(tool_input=query)
        logging.info(f"Search results for '{query}': {results}")
        return results
    except Exception as e:
        logging.error(f"Error searching the web for '{query}': {e}")
        return f"An error occurred while searching the web for '{query}'."
    
@function_tool()    
async def send_email(context: RunContext,  to_email: str, subject: str, message: str, cc_email: Optional[str] = None) -> str:

    """
    Send an email through Gmail.
    
    Args:
        to_email: Recipient email address
        subject: Email subject line
        message: Email body content
        cc_email: Optional CC email address
    """
    try:
        # Gmail SMTP configuration
        smtp_server = "smtp.gmail.com"
        # 587 for submission, often encrypted
        smtp_port = 587
        
        # Get credentials from environment variables
        gmail_user = os.getenv("GMAIL_USER")
        gmail_password = os.getenv("GMAIL_APP_PASSWORD")  # Use App Password, not regular password
        
        if not gmail_user or not gmail_password:
            logging.error("Gmail credentials not found in environment variables")
            return "Email sending failed: Gmail credentials not configured."
        
        # Create message
        msg = MIMEMultipart()
        msg['From'] = gmail_user
        msg['To'] = to_email
        msg['Subject'] = subject
        
        # Add CC if provided
        recipients = [to_email]
        if cc_email:
            msg['Cc'] = cc_email
            recipients.append(cc_email)
        
        # Attach message body
        msg.attach(MIMEText(message, 'plain'))
        
        # Connect to Gmail SMTP server
        server = smtplib.SMTP(smtp_server, smtp_port)
        server.starttls()  # Enable TLS encryption
        server.login(gmail_user, gmail_password)
        
        # Send email
        text = msg.as_string()
        server.sendmail(gmail_user, recipients, text)
        server.quit()
        
        logging.info(f"Email sent successfully to {to_email}")
        return f"Email sent successfully to {to_email}"
        
    except smtplib.SMTPAuthenticationError:
        logging.error("Gmail authentication failed")
        return "Email sending failed: Authentication error. Please check your Gmail credentials."
    except smtplib.SMTPException as e:
        logging.error(f"SMTP error occurred: {e}")
        return f"Email sending failed: SMTP error - {str(e)}"
    except Exception as e:
        logging.error(f"Error sending email: {e}")
        return f"An error occurred while sending email: {str(e)}"
    




FAISS_PATH = "faiss_store"

@function_tool()
async def rag(context: RunContext,   query: str) -> List[Dict[str, Any]]:
    """
    when user ask a question Retrieve relevant documents from faiss store database for a given query as first action automatically.
    
    """
    try:
        logging.info(f"RAG retrieval started for query: {query}")

        
        store = FaissVectorStore(FAISS_PATH)
        store.load()
        # Automatically detect if index there.
        if not os.path.exists(FAISS_PATH):
            logging.info("faiss index not found — building index")
            docs = load_all_documents("data")
            store = FaissVectorStore("faiss_store")
    
        else:
            logging.info("Using existing faiss index")

        
        
        result = store.query(query, top_k=3)
        logging.info(f"RAG retrieval started for query: {result}")
        return result

    except Exception as e:
        logging.error(f"RAG retrieval error for query '{query}': {e}")
        return []
