requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
ddgs
langchain
langchain-core
pypdf
pymupdf
sentence-transformers
chromadb
langchain-groq
jq
docx2txt
faiss-cpu
